Broken tests: Strongly type TestLogContextIdentifier instead of using GUID.
