+++ serilog-utilities-concurrent-correlator/serilog-utilities-concurrent-correlator.csproj
@@ -52,6 +52,9 @@
     <Compile Include="ConcurrentBagSink.cs" />
     <Compile Include="LoggerSinkConfigurationExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestLogContext.cs" />
+    <Compile Include="TestLogContextIdentifier.cs" />
+    <Compile Include="TestSerilogEventsNotConfiguredException.cs" />
     <Compile Include="TestSerilogLogEvents.cs" />
   </ItemGroup>
   <ItemGroup>
